thumbnail generation moved to separate module only if io-write and thumbnail-generation features are both enabled
