+++ threemf2/Cargo.toml
@@ -106,7 +106,7 @@ required-features = ["io-memory-optimized-read"]
 [[example]]
 name = "thumbnail_generation"
 path = "examples/thumbnail_generation.rs"
-required-features = ["thumbnail-generation"]
+required-features = ["io-write", "thumbnail-generation"]
 
 [package.metadata.cargo-all-features]
 max_combination_size = 3
