+++ threemf2/src/lib.rs
@@ -40,3 +40,6 @@ pub mod threemf_namespaces;
     feature = "io-lazy-read"
 ))]
 pub mod io;
+
+#[cfg(all(feature = "io-write", feature = "thumbnail-generation"))]
+pub mod thumbnail;
