+++ threemf2/src/thumbnail/mod.rs
@@ -1,0 +1,4 @@
+mod thumbnail_generator;
+pub use thumbnail_generator::{ThumbnailConfig, ThumbnailGenerator};
+
+mod mesh_pipeline;
